requirements.txt
openai==0.27.5
pandas==2.0.1
textual==0.22.3
python-dotenv~=1.0.0